ggml : stabilize vector index integration
- keep vector index default-off and separate from ggml::all
- split core, search, and persistence implementation files
- fix mmap, delta replay, and compaction edge cases
- tighten SIMD dispatch, search ordering, docs, and tests
